Fixed required changes
