+++ modules/workClass.js
@@ -16,7 +16,7 @@ class TodoList {
       this.id = 0;
       worklist.innerHTML = works.map((work) => {
         if (work.completed === true) {
-          return `<li class="each-todo"><input type="checkbox" class="check_box" id=${work.index}> <p class="text-todo linethrough" id=${work.index}>${work.description}</p><i class="fa-solid fa-trash" id=${work.index}></i></li>`;
+          return `<li class="each-todo"><input type="checkbox" class="check_box" id=${work.index}> <p class="text-todo" id=${work.index}>${work.description}</p><i class="fa-solid fa-trash" id=${work.index}></i></li>`;
         }
         return `<li class="each-todo"><input type="checkbox" class="check_box" id=${work.index}> <p class="text-todo" id=${work.index}>${work.description}</p><i class="fa-solid fa-trash" id=${work.index}></i></li>`;
       }).join('');
@@ -24,6 +24,7 @@ class TodoList {
     } else {
       worklist.innerHTML = 'its empty';
     }
+    this.editTodo();
     interactivelist();
   }
 
@@ -52,5 +53,40 @@ class TodoList {
     localStorage.setItem('todolist', JSON.stringify(search));
     this.workList();
   }
+
+  editedTodos(description, id) {
+    const Todos = JSON.parse(localStorage.getItem('todolist')) || [];
+    if (description === '') {
+      alert('please try something else');
+    } else {
+      const edit = Todos.find((todo) => todo.index.toString() === id);
+      if (edit !== undefined) {
+        edit.description = description;
+        edit.index = id;
+        localStorage.setItem('todolist', JSON.stringify(Todos));
+      }
+      this.workList();
+    }
+  }
+
+  editTodo() {
+    const selected = document.querySelectorAll('.text-todo');
+    for (let i = 0; i < selected.length; i += 1) {
+      selected[i].addEventListener('click', (e) => {
+        const { id } = e.target;
+        const input = document.createElement('input');
+        const inputvalue = selected[i].textContent;
+        selected[i].textContent = '';
+        input.setAttribute('type', 'text');
+        selected[i].appendChild(input);
+        input.focus();
+        input.value = inputvalue;
+        input.addEventListener('focusout', () => {
+          this.editedTodos(input.value, id);
+        });
+      });
+    }
+  }
 }
+
 export default TodoList;
